Enable 96-bit literals only for pnut-exe
Since shell doesn't support integers larger than
32 bits, there's no point in supporting integers
larger than that.
